v0.1.2: opt gui front-ends, needle → search rename, drop notes
v0.1.2-stable: new `opt gui` verb resolves desktop front-ends
(gui_bins + OPTION_GUI_BIN_*), status shows `+ gui (<bin>)`, route
table renames needle → search, and the unshipped notes route is
removed. Metapackage now depends on optionsearch.

+++ src/run.rs
@@ -19,7 +19,7 @@ pub fn run(bin: &PathBuf, args: &[String]) -> Result<ExitCode, String> {
 
     Ok(match status.code() {
         Some(code) => ExitCode::from(code as u8),
-        None => ExitCode::from((128 + signal_number(&status)) as u8),
+        None => ExitCode::from(128 + signal_number(&status)),
     })
 }
